Raises an error when strings are encountered in datasets
A preventative measure was put into place to try and stop breaking NumPy histogram when things with data go wrong. The program will now output an error message if it encounters strings in data. The program will also strip null values from the data before passing it to NumPy histogram